[hh1222] Rename app/firebaseConfig.ts to /firebase.config.ts
